Module Bug Fix [A:0.0.65]
Fixed a bug where modules would not install due to the lack of the
command line. Modules now are installed from the install_dependencies.py
file which is run on startup by D2MT.pyw. Also removed errlog.txt to
avoid confusion that it is a dependency for the app.