Search and Table Display Update [V:0.0.90]
Created a searchable entry for the table which can be used to search for
a specific team, date, time, or tournament. Search box is toggleable
within the custom widget so that you can have a table without a search
box. Also added headers in between columns for simple graphical effect,
which are specified when widget instance is created.